requirements.txt
pyyaml==6.0
selenium==4.1.3